remove additional file
